% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addProbs_vanilla.R
\docType{class}
\name{addedProbs-class}
\alias{addedProbs-class}
\title{\linkS4class{addedProbs}}
\description{
..
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\link[base]{matrix}}

\item{\code{margin}}{\link[base]{integer} scalar or \link[base]{vector}}
}}

