% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binTab.R
\name{as_flextable.binTab}
\alias{as_flextable.binTab}
\title{\code{\link[=as_flextable.binTab]{as_flextable.binTab()}}}
\usage{
\method{as_flextable}{binTab}(x, ...)
}
\arguments{
\item{x}{\code{binTab}}

\item{...}{..}
}
\description{
\code{\link[=as_flextable.binTab]{as_flextable.binTab()}}
}
