% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0S4_Boolean.R
\docType{class}
\name{BooleanRisk-class}
\alias{BooleanRisk-class}
\title{Boolean Risk-&-Disease Table}
\description{
Boolean Risk-&-Disease Table
}
\section{Slots}{

\describe{
\item{\code{.Data}}{two-by-two \link[base]{integer} \link[base]{matrix}, contingency table of a Boolean
test-&-disease decisions with layout
\tabular{lcc}{
 \tab Disease (\eqn{+}) \tab Disease (\eqn{-}) \cr
Risk Factor (\eqn{+}) \tab \eqn{x_{++}} \tab \eqn{x_{+-}} \cr
Risk Factor (\eqn{-}) \tab \eqn{x_{-+}} \tab \eqn{x_{--}} \cr
}}
}}

