% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BooleanTable.R
\docType{class}
\name{BooleanTable-class}
\alias{BooleanTable-class}
\title{Boolean Test-&-Disease and/or Risk-&-Disease Table}
\description{
Boolean Test-&-Disease and/or Risk-&-Disease Table
}
\section{Slots}{

\describe{
\item{\code{.Data}}{two-by-two \link[base]{integer} \link[base]{matrix}, 
contingency table of a Boolean test-&-disease table with layout
\tabular{lcc}{
 \tab Test (\eqn{+}) \tab Test (\eqn{-}) \cr
Disease (\eqn{+}) \tab \eqn{x_{++}} \tab \eqn{x_{+-}} \cr
Disease (\eqn{-}) \tab \eqn{x_{-+}} \tab \eqn{x_{--}} \cr
}
or a Boolean risk-&-disease table with layout
\tabular{lcc}{
 \tab Risk Factor (\eqn{+}) \tab Risk Factor (\eqn{-}) \cr
Disease (\eqn{+}) \tab \eqn{x_{++}} \tab \eqn{x_{+-}} \cr
Disease (\eqn{-}) \tab \eqn{x_{-+}} \tab \eqn{x_{--}} \cr
}
The endpoint (i.e., disease) must be on the rows and the test/risk be on the columns.
This set up is to accommodate \link[stats]{model.frame.default} and let end user use formula
`endpoint ~ test` or `endpoint ~ risk`.}
}}

