% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BooleanTable.R
\name{BooleanTable}
\alias{BooleanTable}
\title{Create Boolean Test-&-Disease and/or Risk-&-Disease Table}
\usage{
BooleanTable(x)
}
\arguments{
\item{x}{two-by-two \link[base]{integer} \link[base]{matrix}, contingency table of two Boolean variables, 
or an R object convertible to a two-by-two \link[base]{integer} \link[base]{matrix}.
The endpoint (i.e., disease) is on rows and the test/risk is on columns.}
}
\value{
[BooleanTable] returns a \linkS4class{BooleanTable} object.
}
\description{
To create Boolean Test-&-Disease and/or Risk-&-Disease table.
}
\details{
..
}
\examples{
x = matrix(c(7L, 3L, 8L, 6L), nrow = 2L)
BooleanTable(x)
(x1 = matrix(c(7L, 3L, 8L, 6L), nrow = 2L, dimnames = list(X = c('a','b'), NULL)))
BooleanTable(x1)

}
\seealso{
End-user may also use \link[caret]{confusionMatrix}, 
which does not provide confidence intervals of 
sensitivity, specificity, etc.
}
