% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0S4_Boolean.R
\docType{class}
\name{BooleanTest-class}
\alias{BooleanTest-class}
\title{Boolean Test-&-Disease Table}
\description{
Boolean Test-&-Disease Table
}
\section{Slots}{

\describe{
\item{\code{.Data}}{two-by-two \link[base]{integer} \link[base]{matrix}, contingency table of a Boolean
test-&-disease decisions with layout
\tabular{lcc}{
 \tab Disease (\eqn{+}) \tab Disease (\eqn{-}) \cr
Test (\eqn{+}) \tab \eqn{x_{++}} \tab \eqn{x_{+-}} \cr
Test (\eqn{-}) \tab \eqn{x_{-+}} \tab \eqn{x_{--}} \cr
}}
}}

