% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chp1.R
\name{Chapter01}
\alias{Chapter01}
\alias{sampleRow}
\title{Chapter 1}
\usage{
sampleRow(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{a \link[base]{data.frame}}

\item{size}{positive \link[base]{integer} scalar, number of rows to be selected}

\item{replace}{\link[base]{logical} scalar, whether sampling should be with replacement (default \code{FALSE})}

\item{prob}{\link[base]{numeric} vector of probability weights for each row of input \code{x} being sampled.
Default \code{NULL} indicates simple random sampling}
}
\value{
\link{sampleRow} returns a \link[base]{data.frame}, a simple random sample from the input.
}
\description{
Functions and examples for Chapter 1, \emph{Introduction to Biostatistics}.
}
\examples{
library(DanielBiostatistics10th)
# To run a line of code, use shortcut
# Command + Enter: Mac and RStudio Cloud
# Control + Enter: Windows, Mac and RStudio Cloud
# To clear the console
# Control + L: Mac and RStudio Cloud

# Page 8, Example 1.4.1
class(EXA_C01_S04_01) # `EXA_C01_S04_01` is a 'data.frame' (a specific class defined in R)
dim(EXA_C01_S04_01) # dimension, number-row and number-column
head(EXA_C01_S04_01, n = 8L) # first `n` rows of a 'data.frame'
names(EXA_C01_S04_01) # column names of a 'data.frame'
EXA_C01_S04_01$AGE # use `$` to obtain one column from a 'data.frame' 
sampleRow(EXA_C01_S04_01, size = 10L, replace = FALSE) # to answer Example 1.4.1

# Page 11, Example 1.4.2
EXA_C01_S04_01[seq.int(from = 4L, to = 166L, by = 18L), ]
}
\references{
Wayne W. Daniel, \emph{Biostatistics: A Foundation for Analysis in the Health Sciences}, Tenth Edition.
Wiley, ISBN: 978-1-119-62550-6.
}
\seealso{
\link[base]{sample.int}
}
