% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{DanielBiostatistics10th-package}
\alias{DanielBiostatistics10th}
\alias{DanielBiostatistics10th-package}
\title{Functions for Wayne W. Daniel's \emph{Biostatistics} (Tenth Edition)}
\description{
Functions and examples to accompany Wayne W. Daniel's
\emph{Biostatistics: A Foundation for Analysis in the Health Sciences}, Tenth Edition,
Wiley, ISBN: 978-1-119-62550-6.

\url{https://www.wiley.com/en-us/Biostatistics:+A+Foundation+for+Analysis+in+the+Health+Sciences,+10th+Edition-p-9781119625506}

Data sets from 10th edition \url{https://bcs.wiley.com/he-bcs/Books?action=resource&bcsId=7849&itemId=1118302796&resourceId=30373}.

Resources from 11th edition \url{https://bcs.wiley.com/he-bcs/Books?action=index&bcsId=11491&itemId=1119496578},
with errata of data.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

}
