% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXA_C01_S04_01}
\alias{EXA_C01_S04_01}
\title{Data of Example 1.4.1}
\format{
A \link[base]{data.frame} with 189 rows and 2 columns
}
\usage{
EXA_C01_S04_01
}
\description{
\code{EXA_C01_S04_01}
}
\keyword{datasets}
