% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXA_C02_S05_05}
\alias{EXA_C02_S05_05}
\title{Data of Example 2.5.5}
\format{
A \link[base]{data.frame} with 20 rows and 1 columns
}
\usage{
EXA_C02_S05_05
}
\description{
\code{EXA_C02_S05_05}
}
\keyword{datasets}
