% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXA_C06_S02_04}
\alias{EXA_C06_S02_04}
\title{Data of Example 6.2.4}
\format{
A \link[base]{data.frame} with 35 rows and 1 columns
}
\usage{
EXA_C06_S02_04
}
\description{
\code{EXA_C06_S02_04}
}
\keyword{datasets}
