% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXA_C07_S02_03}
\alias{EXA_C07_S02_03}
\title{Data of Example 7.2.3}
\format{
A \link[base]{data.frame} with 17 rows and 2 columns
}
\usage{
EXA_C07_S02_03
}
\description{
\code{EXA_C07_S02_03}
}
\keyword{datasets}
