% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXA_C08_S04_01}
\alias{EXA_C08_S04_01}
\title{Data of Example 8.4.1}
\format{
A \link[base]{data.frame} with 72 rows and 3 columns
}
\usage{
EXA_C08_S04_01
}
\description{
\code{EXA_C08_S04_01}
}
\keyword{datasets}
