% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXA_C08_S04_02}
\alias{EXA_C08_S04_02}
\title{Data of Example 8.4.2}
\format{
A \link[base]{data.frame} with 25 rows and 6 columns
}
\usage{
EXA_C08_S04_02
}
\description{
\code{EXA_C08_S04_02}
}
\keyword{datasets}
