% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXA_C11_S01_02}
\alias{EXA_C11_S01_02}
\title{Data of Example 11.1.2}
\format{
A \link[base]{data.frame} with 15 rows and 4 columns
}
\usage{
EXA_C11_S01_02
}
\description{
\code{EXA_C11_S01_02}
}
\keyword{datasets}
