% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXA_C11_S02_01}
\alias{EXA_C11_S02_01}
\title{Data of Example 11.2.1}
\format{
A \link[base]{data.frame} with 100 rows and 4 columns
}
\usage{
EXA_C11_S02_01
}
\description{
\code{EXA_C11_S02_01}
}
\keyword{datasets}
