% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXA_C11_S02_03}
\alias{EXA_C11_S02_03}
\title{Data of Example 11.2.3}
\format{
A \link[base]{data.frame} with 36 rows and 3 columns
}
\usage{
EXA_C11_S02_03
}
\description{
\code{EXA_C11_S02_03}
}
\keyword{datasets}
