% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXA_C13_S07_01}
\alias{EXA_C13_S07_01}
\title{Data of Example 13.7.1}
\format{
A \link[base]{data.frame} with 36 rows and 1 columns
}
\usage{
EXA_C13_S07_01
}
\description{
\code{EXA_C13_S07_01}
}
\keyword{datasets}
