% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXA_C13_S09_02}
\alias{EXA_C13_S09_02}
\title{Data of Example 13.9.2}
\format{
A \link[base]{data.frame} with 16 rows and 5 columns
}
\usage{
EXA_C13_S09_02
}
\description{
\code{EXA_C13_S09_02}
}
\keyword{datasets}
