% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C02_S03_01}
\alias{EXR_C02_S03_01}
\title{Data of Exercise 2.3.1}
\format{
A \link[base]{data.frame} with 90 rows and 1 columns
}
\usage{
EXR_C02_S03_01
}
\description{
\code{EXR_C02_S03_01}
}
\keyword{datasets}
