% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C02_S03_04}
\alias{EXR_C02_S03_04}
\title{Data of Exercise 2.3.4}
\format{
A \link[base]{data.frame} with 53 rows and 1 columns
}
\usage{
EXR_C02_S03_04
}
\description{
\code{EXR_C02_S03_04}
}
\keyword{datasets}
