% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C02_S03_06}
\alias{EXR_C02_S03_06}
\title{Data of Exercise 2.3.6}
\format{
A \link[base]{data.frame} with 60 rows and 1 columns
}
\usage{
EXR_C02_S03_06
}
\description{
\code{EXR_C02_S03_06}
}
\keyword{datasets}
