% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C02_S03_07}
\alias{EXR_C02_S03_07}
\title{Data of Exercise 2.3.7}
\format{
A \link[base]{data.frame} with 155 rows and 1 columns
}
\usage{
EXR_C02_S03_07
}
\description{
\code{EXR_C02_S03_07}
}
\keyword{datasets}
