% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C02_S03_09}
\alias{EXR_C02_S03_09}
\title{Data of Exercise 2.3.9}
\format{
A \link[base]{data.frame} with 50 rows and 2 columns
}
\usage{
EXR_C02_S03_09
}
\description{
\code{EXR_C02_S03_09}
}
\keyword{datasets}
