% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C02_S03_11}
\alias{EXR_C02_S03_11}
\title{Data of Exercise 2.3.11}
\format{
A \link[base]{data.frame} with 216 rows and 1 columns
}
\usage{
EXR_C02_S03_11
}
\description{
\code{EXR_C02_S03_11}
}
\keyword{datasets}
