% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C02_S05_03}
\alias{EXR_C02_S05_03}
\title{Data of Exercise 2.5.3}
\format{
A \link[base]{data.frame} with 30 rows and 1 columns
}
\usage{
EXR_C02_S05_03
}
\description{
\code{EXR_C02_S05_03}
}
\keyword{datasets}
