% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C06_S02_05}
\alias{EXR_C06_S02_05}
\title{Data of Exercise 6.2.5}
\format{
A \link[base]{data.frame} with 16 rows and 1 columns
}
\usage{
EXR_C06_S02_05
}
\description{
\code{EXR_C06_S02_05}
}
\keyword{datasets}
