% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C06_S04_10}
\alias{EXR_C06_S04_10}
\title{Data of Exercise 6.4.10}
\format{
A \link[base]{data.frame} with 32 rows and 2 columns
}
\usage{
EXR_C06_S04_10
}
\description{
\code{EXR_C06_S04_10}
}
\keyword{datasets}
