% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C06_S10_07}
\alias{EXR_C06_S10_07}
\title{Data of Exercise 6.10.7}
\format{
A \link[base]{data.frame} with 26 rows and 2 columns
}
\usage{
EXR_C06_S10_07
}
\description{
\code{EXR_C06_S10_07}
}
\keyword{datasets}
