% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C07_S02_15}
\alias{EXR_C07_S02_15}
\title{Data of Exercise 7.2.15}
\format{
A \link[base]{data.frame} with 50 rows and 1 columns
}
\usage{
EXR_C07_S02_15
}
\description{
\code{EXR_C07_S02_15}
}
\keyword{datasets}
