% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C07_S02_16}
\alias{EXR_C07_S02_16}
\title{Data of Exercise 7.2.16}
\format{
A \link[base]{data.frame} with 21 rows and 1 columns
}
\usage{
EXR_C07_S02_16
}
\description{
\code{EXR_C07_S02_16}
}
\keyword{datasets}
