% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C07_S03_05}
\alias{EXR_C07_S03_05}
\title{Data of Exercise 7.3.5}
\format{
A \link[base]{data.frame} with 82 rows and 2 columns
}
\usage{
EXR_C07_S03_05
}
\description{
\code{EXR_C07_S03_05}
}
\keyword{datasets}
