% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C07_S03_10}
\alias{EXR_C07_S03_10}
\title{Data of Exercise 7.3.10}
\format{
A \link[base]{data.frame} with 24 rows and 2 columns
}
\usage{
EXR_C07_S03_10
}
\description{
\code{EXR_C07_S03_10}
}
\keyword{datasets}
