% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C07_S08_07}
\alias{EXR_C07_S08_07}
\title{Data of Exercise 7.8.7}
\format{
A \link[base]{data.frame} with 23 rows and 2 columns
}
\usage{
EXR_C07_S08_07
}
\description{
\code{EXR_C07_S08_07}
}
\keyword{datasets}
