% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C08_S02_05}
\alias{EXR_C08_S02_05}
\title{Data of Exercise 8.2.5}
\format{
A \link[base]{data.frame} with 29 rows and 2 columns
}
\usage{
EXR_C08_S02_05
}
\description{
\code{EXR_C08_S02_05}
}
\keyword{datasets}
