% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C08_S02_06}
\alias{EXR_C08_S02_06}
\title{Data of Exercise 8.2.6}
\format{
A \link[base]{data.frame} with 90 rows and 2 columns
}
\usage{
EXR_C08_S02_06
}
\description{
\code{EXR_C08_S02_06}
}
\keyword{datasets}
