% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C08_S04_03}
\alias{EXR_C08_S04_03}
\title{Data of Exercise 8.4.3}
\format{
A \link[base]{data.frame} with 48 rows and 3 columns
}
\usage{
EXR_C08_S04_03
}
\description{
\code{EXR_C08_S04_03}
}
\keyword{datasets}
