% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C08_S04_06}
\alias{EXR_C08_S04_06}
\title{Data of Exercise 8.4.6}
\format{
A \link[base]{data.frame} with 20 rows and 6 columns
}
\usage{
EXR_C08_S04_06
}
\description{
\code{EXR_C08_S04_06}
}
\keyword{datasets}
