% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C08_S05_01}
\alias{EXR_C08_S05_01}
\title{Data of Exercise 8.5.1}
\format{
A \link[base]{data.frame} with 24 rows and 3 columns
}
\usage{
EXR_C08_S05_01
}
\description{
\code{EXR_C08_S05_01}
}
\keyword{datasets}
