% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C09_S03_07}
\alias{EXR_C09_S03_07}
\title{Data of Exercise 9.3.7}
\format{
A \link[base]{data.frame} with 27 rows and 2 columns
}
\usage{
EXR_C09_S03_07
}
\description{
\code{EXR_C09_S03_07}
}
\keyword{datasets}
