% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C09_S07_04}
\alias{EXR_C09_S07_04}
\title{Data of Exercise 9.7.4}
\format{
A \link[base]{data.frame} with 18 rows and 2 columns
}
\usage{
EXR_C09_S07_04
}
\description{
\code{EXR_C09_S07_04}
}
\keyword{datasets}
