% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C09_S07_06}
\alias{EXR_C09_S07_06}
\title{Data of Exercise 9.7.6}
\format{
A \link[base]{data.frame} with 15 rows and 2 columns
}
\usage{
EXR_C09_S07_06
}
\description{
\code{EXR_C09_S07_06}
}
\keyword{datasets}
