% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C10_S03_01}
\alias{EXR_C10_S03_01}
\title{Data of Exercise 10.3.1}
\format{
A \link[base]{data.frame} with 35 rows and 3 columns
}
\usage{
EXR_C10_S03_01
}
\description{
\code{EXR_C10_S03_01}
}
\keyword{datasets}
