% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C10_S06_02}
\alias{EXR_C10_S06_02}
\title{Data of Exercise 10.6.2}
\format{
A \link[base]{data.frame} with 12 rows and 3 columns
}
\usage{
EXR_C10_S06_02
}
\description{
\code{EXR_C10_S06_02}
}
\keyword{datasets}
