% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C10_S06_04}
\alias{EXR_C10_S06_04}
\title{Data of Exercise 10.6.4}
\format{
A \link[base]{data.frame} with 15 rows and 6 columns
}
\usage{
EXR_C10_S06_04
}
\description{
\code{EXR_C10_S06_04}
}
\keyword{datasets}
