% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C11_S02_04}
\alias{EXR_C11_S02_04}
\title{Data of Exercise 11.2.4}
\format{
A \link[base]{data.frame} with 90 rows and 3 columns
}
\usage{
EXR_C11_S02_04
}
\description{
\code{EXR_C11_S02_04}
}
\keyword{datasets}
