% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C11_S03_03}
\alias{EXR_C11_S03_03}
\title{Data of Exercise 11.3.3}
\format{
A \link[base]{data.frame} with 68 rows and 7 columns
}
\usage{
EXR_C11_S03_03
}
\description{
\code{EXR_C11_S03_03}
}
\keyword{datasets}
