% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C11_S04_01}
\alias{EXR_C11_S04_01}
\title{Data of Exercise 11.4.1}
\format{
A \link[base]{data.frame} with 4 rows and 3 columns
}
\usage{
EXR_C11_S04_01
}
\description{
\code{EXR_C11_S04_01}
}
\keyword{datasets}
