% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C11_S05_01}
\alias{EXR_C11_S05_01}
\title{Data of Exercise 11.5.1}
\format{
A \link[base]{data.frame} with 45 rows and 4 columns
}
\usage{
EXR_C11_S05_01
}
\description{
\code{EXR_C11_S05_01}
}
\keyword{datasets}
