% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C13_S07_02}
\alias{EXR_C13_S07_02}
\title{Data of Exercise 13.7.2}
\format{
A \link[base]{data.frame} with 30 rows and 1 columns
}
\usage{
EXR_C13_S07_02
}
\description{
\code{EXR_C13_S07_02}
}
\keyword{datasets}
