% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C13_S08_03}
\alias{EXR_C13_S08_03}
\title{Data of Exercise 13.8.3}
\format{
A \link[base]{data.frame} with 53 rows and 2 columns
}
\usage{
EXR_C13_S08_03
}
\description{
\code{EXR_C13_S08_03}
}
\keyword{datasets}
