% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C13_S08_04}
\alias{EXR_C13_S08_04}
\title{Data of Exercise 13.8.4}
\format{
A \link[base]{data.frame} with 22 rows and 2 columns
}
\usage{
EXR_C13_S08_04
}
\description{
\code{EXR_C13_S08_04}
}
\keyword{datasets}
