% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C13_S08_06}
\alias{EXR_C13_S08_06}
\title{Data of Exercise 13.8.6}
\format{
A \link[base]{data.frame} with 22 rows and 3 columns
}
\usage{
EXR_C13_S08_06
}
\description{
\code{EXR_C13_S08_06}
}
\keyword{datasets}
