% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C13_S09_03}
\alias{EXR_C13_S09_03}
\title{Data of Exercise 13.9.3}
\format{
A \link[base]{data.frame} with 10 rows and 6 columns
}
\usage{
EXR_C13_S09_03
}
\description{
\code{EXR_C13_S09_03}
}
\keyword{datasets}
