% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C13_S10_01}
\alias{EXR_C13_S10_01}
\title{Data of Exercise 13.10.1}
\format{
A \link[base]{data.frame} with 15 rows and 3 columns
}
\usage{
EXR_C13_S10_01
}
\description{
\code{EXR_C13_S10_01}
}
\keyword{datasets}
