% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C13_S10_02}
\alias{EXR_C13_S10_02}
\title{Data of Exercise 13.10.2}
\format{
A \link[base]{data.frame} with 10 rows and 3 columns
}
\usage{
EXR_C13_S10_02
}
\description{
\code{EXR_C13_S10_02}
}
\keyword{datasets}
