% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C13_S10_05}
\alias{EXR_C13_S10_05}
\title{Data of Exercise 13.10.5}
\format{
A \link[base]{data.frame} with 30 rows and 2 columns
}
\usage{
EXR_C13_S10_05
}
\description{
\code{EXR_C13_S10_05}
}
\keyword{datasets}
