% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{EXR_C14_S04_03}
\alias{EXR_C14_S04_03}
\title{Data of Exercise 14.4.3}
\format{
A \link[base]{data.frame} with 50 rows and 4 columns
}
\usage{
EXR_C14_S04_03
}
\description{
\code{EXR_C14_S04_03}
}
\keyword{datasets}
