% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chp2.R
\docType{class}
\name{freqs-class}
\alias{freqs-class}
\title{S4 Class \linkS4class{freqs}}
\description{
S4 Class \linkS4class{freqs}
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\link[base]{integer} \link[base]{vector}, frequency counts}

\item{\code{data.name}}{\link[base]{character} integer, name of the data, only used in output}
}}

\keyword{internal}
