% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDS_doc.R
\docType{data}
\name{LDS}
\alias{LDS}
\alias{LDS_C02_NCBIRTH800}
\alias{LDS_C06_BABYWGTS}
\alias{LDS_C06_BOYHGTS}
\alias{LDS_C06_CHOLEST}
\alias{LDS_C07_HEADCIRC}
\alias{LDS_C07_HEMOGLOB}
\alias{LDS_C07_MANDEXT}
\alias{LDS_C07_PCKDATA}
\alias{LDS_C07_PROTHROM}
\alias{LDS_C08_CSFDATA}
\alias{LDS_C08_LSADATA}
\alias{LDS_C08_MEDSCORES}
\alias{LDS_C08_RBCDATA}
\alias{LDS_C08_SACEDATA}
\alias{LDS_C08_SERUMCHO}
\alias{LDS_C09_CALCIUM}
\alias{LDS_C09_CEREBRAL}
\alias{LDS_C09_HYPERTEN}
\alias{LDS_C10_LTEXER}
\alias{LDS_C10_RESPDIS}
\alias{LDS_C10_RISKFACT}
\alias{LDS_C10_STERLENGTH}
\alias{LDS_C11_AQUATICS}
\alias{LDS_C11_TEACHERS}
\alias{LDS_C11_WGTLOSS}
\alias{LDS_C12_SMOKING}
\alias{LDS_C13_KLETTER}
\title{LDS Data}
\format{
An object of class \code{data.frame} with 800 rows and 14 columns.

An object of class \code{data.frame} with 1200 rows and 2 columns.

An object of class \code{data.frame} with 1000 rows and 2 columns.

An object of class \code{data.frame} with 1000 rows and 2 columns.

An object of class \code{data.frame} with 1000 rows and 3 columns.

An object of class \code{data.frame} with 1000 rows and 2 columns.

An object of class \code{data.frame} with 1000 rows and 2 columns.

An object of class \code{data.frame} with 1005 rows and 3 columns.

An object of class \code{data.frame} with 1000 rows and 2 columns.

An object of class \code{data.frame} with 300 rows and 6 columns.

An object of class \code{data.frame} with 350 rows and 5 columns.

An object of class \code{data.frame} with 582 rows and 4 columns.

An object of class \code{data.frame} with 350 rows and 4 columns.

An object of class \code{data.frame} with 400 rows and 5 columns.

An object of class \code{data.frame} with 347 rows and 4 columns.

An object of class \code{data.frame} with 100 rows and 13 columns.

An object of class \code{data.frame} with 1050 rows and 3 columns.

An object of class \code{data.frame} with 1050 rows and 3 columns.

An object of class \code{data.frame} with 248 rows and 5 columns.

An object of class \code{data.frame} with 1200 rows and 6 columns.

An object of class \code{data.frame} with 1000 rows and 6 columns.

An object of class \code{data.frame} with 1162 rows and 4 columns.

An object of class \code{data.frame} with 142 rows and 7 columns.

An object of class \code{data.frame} with 212 rows and 7 columns.

An object of class \code{data.frame} with 1185 rows and 3 columns.

An object of class \code{data.frame} with 1200 rows and 6 columns.

An object of class \code{data.frame} with 168 rows and 3 columns.
}
\usage{
LDS_C02_NCBIRTH800

LDS_C06_BABYWGTS

LDS_C06_BOYHGTS

LDS_C06_CHOLEST

LDS_C07_HEADCIRC

LDS_C07_HEMOGLOB

LDS_C07_MANDEXT

LDS_C07_PCKDATA

LDS_C07_PROTHROM

LDS_C08_CSFDATA

LDS_C08_LSADATA

LDS_C08_MEDSCORES

LDS_C08_RBCDATA

LDS_C08_SACEDATA

LDS_C08_SERUMCHO

LDS_C09_CALCIUM

LDS_C09_CEREBRAL

LDS_C09_HYPERTEN

LDS_C10_LTEXER

LDS_C10_RESPDIS

LDS_C10_RISKFACT

LDS_C10_STERLENGTH

LDS_C11_AQUATICS

LDS_C11_TEACHERS

LDS_C11_WGTLOSS

LDS_C12_SMOKING

LDS_C13_KLETTER
}
\description{
LDS Data
}
\keyword{datasets}
\keyword{internal}
