% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C02_NCBIRTH800}
\alias{LDS_C02_NCBIRTH800}
\title{Large Data \code{NCBIRTH800} from Chapter 2}
\format{
A \link[base]{data.frame} with 800 rows and 14 columns
}
\usage{
LDS_C02_NCBIRTH800
}
\description{
\code{LDS_C02_NCBIRTH800}
}
\keyword{datasets}
