% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C06_CHOLEST}
\alias{LDS_C06_CHOLEST}
\title{Large Data \code{CHOLEST} from Chapter 6}
\format{
A \link[base]{data.frame} with 1000 rows and 2 columns
}
\usage{
LDS_C06_CHOLEST
}
\description{
\code{LDS_C06_CHOLEST}
}
\keyword{datasets}
