% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C07_HEADCIRC}
\alias{LDS_C07_HEADCIRC}
\title{Large Data \code{HEADCIRC} from Chapter 7}
\format{
A \link[base]{data.frame} with 1000 rows and 3 columns
}
\usage{
LDS_C07_HEADCIRC
}
\description{
\code{LDS_C07_HEADCIRC}
}
\keyword{datasets}
