% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C07_HEMOGLOB}
\alias{LDS_C07_HEMOGLOB}
\title{Large Data \code{HEMOGLOB} from Chapter 7}
\format{
A \link[base]{data.frame} with 1000 rows and 2 columns
}
\usage{
LDS_C07_HEMOGLOB
}
\description{
\code{LDS_C07_HEMOGLOB}
}
\keyword{datasets}
