% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C07_MANDEXT}
\alias{LDS_C07_MANDEXT}
\title{Large Data \code{MANDEXT} from Chapter 7}
\format{
A \link[base]{data.frame} with 1000 rows and 2 columns
}
\usage{
LDS_C07_MANDEXT
}
\description{
\code{LDS_C07_MANDEXT}
}
\keyword{datasets}
