% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C07_PCKDATA}
\alias{LDS_C07_PCKDATA}
\title{Large Data \code{PCKDATA} from Chapter 7}
\format{
A \link[base]{data.frame} with 1005 rows and 3 columns
}
\usage{
LDS_C07_PCKDATA
}
\description{
\code{LDS_C07_PCKDATA}
}
\keyword{datasets}
