% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C08_CSFDATA}
\alias{LDS_C08_CSFDATA}
\title{Large Data \code{CSFDATA} from Chapter 8}
\format{
A \link[base]{data.frame} with 300 rows and 6 columns
}
\usage{
LDS_C08_CSFDATA
}
\description{
\code{LDS_C08_CSFDATA}
}
\keyword{datasets}
