% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C08_LSADATA}
\alias{LDS_C08_LSADATA}
\title{Large Data \code{LSADATA} from Chapter 8}
\format{
A \link[base]{data.frame} with 350 rows and 5 columns
}
\usage{
LDS_C08_LSADATA
}
\description{
\code{LDS_C08_LSADATA}
}
\keyword{datasets}
