% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C08_RBCDATA}
\alias{LDS_C08_RBCDATA}
\title{Large Data \code{RBCDATA} from Chapter 8}
\format{
A \link[base]{data.frame} with 350 rows and 4 columns
}
\usage{
LDS_C08_RBCDATA
}
\description{
\code{LDS_C08_RBCDATA}
}
\keyword{datasets}
