% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C08_SACEDATA}
\alias{LDS_C08_SACEDATA}
\title{Large Data \code{SACEDATA} from Chapter 8}
\format{
A \link[base]{data.frame} with 400 rows and 5 columns
}
\usage{
LDS_C08_SACEDATA
}
\description{
\code{LDS_C08_SACEDATA}
}
\keyword{datasets}
