% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C08_SERUMCHO}
\alias{LDS_C08_SERUMCHO}
\title{Large Data \code{SERUMCHO} from Chapter 8}
\format{
A \link[base]{data.frame} with 347 rows and 4 columns
}
\usage{
LDS_C08_SERUMCHO
}
\description{
\code{LDS_C08_SERUMCHO}
}
\keyword{datasets}
