% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C09_CALCIUM}
\alias{LDS_C09_CALCIUM}
\title{Large Data \code{CALCIUM} from Chapter 9}
\format{
A \link[base]{data.frame} with 100 rows and 13 columns
}
\usage{
LDS_C09_CALCIUM
}
\description{
\code{LDS_C09_CALCIUM}
}
\keyword{datasets}
