% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C09_CEREBRAL}
\alias{LDS_C09_CEREBRAL}
\title{Large Data \code{CEREBRAL} from Chapter 9}
\format{
A \link[base]{data.frame} with 1050 rows and 3 columns
}
\usage{
LDS_C09_CEREBRAL
}
\description{
\code{LDS_C09_CEREBRAL}
}
\keyword{datasets}
