% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C10_LTEXER}
\alias{LDS_C10_LTEXER}
\title{Large Data \code{LTEXER} from Chapter 10}
\format{
A \link[base]{data.frame} with 248 rows and 5 columns
}
\usage{
LDS_C10_LTEXER
}
\description{
\code{LDS_C10_LTEXER}
}
\keyword{datasets}
