% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C10_RISKFACT}
\alias{LDS_C10_RISKFACT}
\title{Large Data \code{RISKFACT} from Chapter 10}
\format{
A \link[base]{data.frame} with 1000 rows and 6 columns
}
\usage{
LDS_C10_RISKFACT
}
\description{
\code{LDS_C10_RISKFACT}
}
\keyword{datasets}
