% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C10_STERLENGTH}
\alias{LDS_C10_STERLENGTH}
\title{Large Data \code{STERLENGTH} from Chapter 10}
\format{
A \link[base]{data.frame} with 1162 rows and 4 columns
}
\usage{
LDS_C10_STERLENGTH
}
\description{
\code{LDS_C10_STERLENGTH}
}
\keyword{datasets}
