% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C11_AQUATICS}
\alias{LDS_C11_AQUATICS}
\title{Large Data \code{AQUATICS} from Chapter 11}
\format{
A \link[base]{data.frame} with 142 rows and 7 columns
}
\usage{
LDS_C11_AQUATICS
}
\description{
\code{LDS_C11_AQUATICS}
}
\keyword{datasets}
