% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C11_TEACHERS}
\alias{LDS_C11_TEACHERS}
\title{Large Data \code{TEACHERS} from Chapter 11}
\format{
A \link[base]{data.frame} with 212 rows and 7 columns
}
\usage{
LDS_C11_TEACHERS
}
\description{
\code{LDS_C11_TEACHERS}
}
\keyword{datasets}
