% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C11_WGTLOSS}
\alias{LDS_C11_WGTLOSS}
\title{Large Data \code{WGTLOSS} from Chapter 11}
\format{
A \link[base]{data.frame} with 1185 rows and 3 columns
}
\usage{
LDS_C11_WGTLOSS
}
\description{
\code{LDS_C11_WGTLOSS}
}
\keyword{datasets}
