% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C12_SMOKING}
\alias{LDS_C12_SMOKING}
\title{Large Data \code{SMOKING} from Chapter 12}
\format{
A \link[base]{data.frame} with 1200 rows and 6 columns
}
\usage{
LDS_C12_SMOKING
}
\description{
\code{LDS_C12_SMOKING}
}
\keyword{datasets}
