% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LDS_C13_KLETTER}
\alias{LDS_C13_KLETTER}
\title{Large Data \code{KLETTER} from Chapter 13}
\format{
A \link[base]{data.frame} with 168 rows and 3 columns
}
\usage{
LDS_C13_KLETTER
}
\description{
\code{LDS_C13_KLETTER}
}
\keyword{datasets}
