% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chp7.power.R
\docType{class}
\name{power-class}
\alias{power-class}
\title{S4 class \linkS4class{power}}
\description{
S4 class \linkS4class{power}
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\link[base]{numeric} scalar or vector, power(s) calculated at
alternative parameter(s) \eqn{\mu_1}}

\item{\code{x}}{\link[base]{numeric} scalar or vector, alternative parameter(s) \eqn{\mu_1}}

\item{\code{rr}}{\linkS4class{RejectionRegion} object}
}}

