% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binTab.R
\name{predval}
\alias{predval}
\alias{ppv}
\alias{npv}
\title{Predictive Values}
\usage{
ppv(prevalence, sensitivity, specificity)

npv(prevalence, sensitivity, specificity)
}
\arguments{
\item{prevalence}{\link[base]{double} scalar or \link[base]{vector}}

\item{sensitivity, specificity}{\link[base]{double} scalars}
}
\value{
Functions \link{ppv} and \link{npv} return \link[base]{double} scalar or \link[base]{vector}.
}
\description{
Positive and negative predictive values
}
\details{
Function \link{ppv} calculates positive predictive values based on input \code{prevalence}.

Function \link{npv} calculates negative predictive values based on input \code{prevalence}.
}
