% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chp2.R
\name{print_stats}
\alias{print_stats}
\alias{print_freqs}
\title{Chapter 2: Descriptive Statistics}
\usage{
print_stats(x, na.rm = TRUE)

print_freqs(x, breaks, include.lowest = TRUE, right = TRUE)
}
\arguments{
\item{x}{\link[base]{numeric} vector, the observations.
In function \link{print_freqs}, this argument can also be a \link[base]{factor}}

\item{na.rm}{\link[base]{logical} scalar, whether to remove the missing observations (default \code{TRUE})}

\item{breaks}{\link[base]{numeric} vector, see \link[base]{cut.default}}

\item{include.lowest}{\link[base]{logical} scalar, default \code{TRUE}. See \link[base]{cut.default}}

\item{right}{\link[base]{logical} scalar, see \link[base]{cut.default}}
}
\value{
Function \link{print_freqs} returns a \linkS4class{freqs} object, for which
a \link[methods]{show} method is defined.

Function \link{print_stats} does not have a returned value.
}
\description{
Functions and examples for Chapter 2, \emph{Descriptive Statistics}.
}
\details{
Function \link{print_freqs} prints the (relative) frequencies and cumulative (relative) frequencies, from
a numeric input vector, specified interval breaks as well as open/close status of the ends of the intervals.

Function \link{print_stats} prints the simple statistics of the input observations, such as sample size,
mean, median, (smallest) mode, variance, standard deviation,
coefficient of variation (if all observations are non-negative),
quartiles, inter-quartile range (IQR), range, skewness and kurtosis.  A histogram is also printed.
}
\seealso{
\link[base]{cut.default} \link[base]{table} \link[base]{cumsum}
\link[base]{mean.default} \link[stats]{median.default} \link[pracma]{Mode}
\link[stats]{var} \link[stats]{sd} \link[stats]{quantile}
\link[e1071]{skewness} \link[e1071]{kurtosis}
}
