% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chp7.power.R
\docType{class}
\name{RejectionRegion-class}
\alias{RejectionRegion-class}
\title{S4 class \linkS4class{RejectionRegion}}
\description{
S4 class \linkS4class{RejectionRegion}
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\link[base]{numeric} scalar for one-sided test, 
or length-two \link[base]{numeric} vector for two-sided test}

\item{\code{null.value}}{\link[base]{numeric} scalar, null value}

\item{\code{std.err}}{\link[base]{numeric} scalar, standard error of sampling distribution}

\item{\code{parameter}}{\link[base]{numeric} vector, additional parameters}

\item{\code{alternative}}{\link[base]{character} scalar, alternative hypothesis}

\item{\code{sig.level}}{\link[base]{numeric} scalar, significance level (Type I error probability)}

\item{\code{test}}{\link[base]{character} scalar, type of test. Currently only \code{'z'} is supported}
}}

