% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C02_13}
\alias{REV_C02_13}
\title{Review Exercise 13 of Chapter 2}
\format{
A \link[base]{data.frame} with 50 rows and 1 columns
}
\usage{
REV_C02_13
}
\description{
\code{REV_C02_13}
}
\keyword{datasets}
