% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C02_16}
\alias{REV_C02_16}
\title{Review Exercise 16 of Chapter 2}
\format{
A \link[base]{data.frame} with 53 rows and 2 columns
}
\usage{
REV_C02_16
}
\description{
\code{REV_C02_16}
}
\keyword{datasets}
