% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C02_19}
\alias{REV_C02_19}
\title{Review Exercise 19 of Chapter 2}
\format{
A \link[base]{data.frame} with 22 rows and 1 columns
}
\usage{
REV_C02_19
}
\description{
\code{REV_C02_19}
}
\keyword{datasets}
