% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C06_23}
\alias{REV_C06_23}
\title{Review Exercise 23 of Chapter 6}
\format{
A \link[base]{data.frame} with 28 rows and 2 columns
}
\usage{
REV_C06_23
}
\description{
\code{REV_C06_23}
}
\keyword{datasets}
