% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C06_28}
\alias{REV_C06_28}
\title{Review Exercise 28 of Chapter 6}
\format{
A \link[base]{data.frame} with 110 rows and 2 columns
}
\usage{
REV_C06_28
}
\description{
\code{REV_C06_28}
}
\keyword{datasets}
