% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C07_19}
\alias{REV_C07_19}
\title{Review Exercise 19 of Chapter 7}
\format{
A \link[base]{data.frame} with 107 rows and 3 columns
}
\usage{
REV_C07_19
}
\description{
\code{REV_C07_19}
}
\keyword{datasets}
