% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C07_22}
\alias{REV_C07_22}
\title{Review Exercise 22 of Chapter 7}
\format{
A \link[base]{data.frame} with 76 rows and 2 columns
}
\usage{
REV_C07_22
}
\description{
\code{REV_C07_22}
}
\keyword{datasets}
