% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C07_41}
\alias{REV_C07_41}
\title{Review Exercise 41 of Chapter 7}
\format{
A \link[base]{data.frame} with 11 rows and 3 columns
}
\usage{
REV_C07_41
}
\description{
\code{REV_C07_41}
}
\keyword{datasets}
