% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C07_42}
\alias{REV_C07_42}
\title{Review Exercise 42 of Chapter 7}
\format{
A \link[base]{data.frame} with 10 rows and 6 columns
}
\usage{
REV_C07_42
}
\description{
\code{REV_C07_42}
}
\keyword{datasets}
