% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C07_44}
\alias{REV_C07_44}
\title{Review Exercise 44 of Chapter 7}
\format{
A \link[base]{data.frame} with 98 rows and 4 columns
}
\usage{
REV_C07_44
}
\description{
\code{REV_C07_44}
}
\keyword{datasets}
