% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C07_45}
\alias{REV_C07_45}
\title{Review Exercise 45 of Chapter 7}
\format{
A \link[base]{data.frame} with 15 rows and 11 columns
}
\usage{
REV_C07_45
}
\description{
\code{REV_C07_45}
}
\keyword{datasets}
