% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C07_46}
\alias{REV_C07_46}
\title{Review Exercise 46 of Chapter 7}
\format{
A \link[base]{data.frame} with 17 rows and 2 columns
}
\usage{
REV_C07_46
}
\description{
\code{REV_C07_46}
}
\keyword{datasets}
