% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C07_47}
\alias{REV_C07_47}
\title{Review Exercise 47 of Chapter 7}
\format{
A \link[base]{data.frame} with 66 rows and 2 columns
}
\usage{
REV_C07_47
}
\description{
\code{REV_C07_47}
}
\keyword{datasets}
