% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C07_50}
\alias{REV_C07_50}
\title{Review Exercise 50 of Chapter 7}
\format{
A \link[base]{data.frame} with 28 rows and 4 columns
}
\usage{
REV_C07_50
}
\description{
\code{REV_C07_50}
}
\keyword{datasets}
