% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_16}
\alias{REV_C08_16}
\title{Review Exercise 16 of Chapter 8}
\format{
A \link[base]{data.frame} with 36 rows and 3 columns
}
\usage{
REV_C08_16
}
\description{
\code{REV_C08_16}
}
\keyword{datasets}
