% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_18}
\alias{REV_C08_18}
\title{Review Exercise 18 of Chapter 8}
\format{
A \link[base]{data.frame} with 74 rows and 4 columns
}
\usage{
REV_C08_18
}
\description{
\code{REV_C08_18}
}
\keyword{datasets}
