% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_19}
\alias{REV_C08_19}
\title{Review Exercise 19 of Chapter 8}
\format{
A \link[base]{data.frame} with 162 rows and 3 columns
}
\usage{
REV_C08_19
}
\description{
\code{REV_C08_19}
}
\keyword{datasets}
