% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_31}
\alias{REV_C08_31}
\title{Review Exercise 31 of Chapter 8}
\format{
A \link[base]{data.frame} with 28 rows and 1 columns
}
\usage{
REV_C08_31
}
\description{
\code{REV_C08_31}
}
\keyword{datasets}
