% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_32}
\alias{REV_C08_32}
\title{Review Exercise 32 of Chapter 8}
\format{
A \link[base]{data.frame} with 48 rows and 1 columns
}
\usage{
REV_C08_32
}
\description{
\code{REV_C08_32}
}
\keyword{datasets}
