% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_33}
\alias{REV_C08_33}
\title{Review Exercise 33 of Chapter 8}
\format{
A \link[base]{data.frame} with 25 rows and 2 columns
}
\usage{
REV_C08_33
}
\description{
\code{REV_C08_33}
}
\keyword{datasets}
