% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_39}
\alias{REV_C08_39}
\title{Review Exercise 39 of Chapter 8}
\format{
A \link[base]{data.frame} with 18 rows and 10 columns
}
\usage{
REV_C08_39
}
\description{
\code{REV_C08_39}
}
\keyword{datasets}
