% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_41}
\alias{REV_C08_41}
\title{Review Exercise 41 of Chapter 8}
\format{
A \link[base]{data.frame} with 24 rows and 2 columns
}
\usage{
REV_C08_41
}
\description{
\code{REV_C08_41}
}
\keyword{datasets}
