% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_42}
\alias{REV_C08_42}
\title{Review Exercise 42 of Chapter 8}
\format{
A \link[base]{data.frame} with 14 rows and 8 columns
}
\usage{
REV_C08_42
}
\description{
\code{REV_C08_42}
}
\keyword{datasets}
