% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_43}
\alias{REV_C08_43}
\title{Review Exercise 43 of Chapter 8}
\format{
A \link[base]{data.frame} with 24 rows and 5 columns
}
\usage{
REV_C08_43
}
\description{
\code{REV_C08_43}
}
\keyword{datasets}
