% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_50}
\alias{REV_C08_50}
\title{Review Exercise 50 of Chapter 8}
\format{
A \link[base]{data.frame} with 20 rows and 5 columns
}
\usage{
REV_C08_50
}
\description{
\code{REV_C08_50}
}
\keyword{datasets}
