% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_53}
\alias{REV_C08_53}
\title{Review Exercise 53 of Chapter 8}
\format{
A \link[base]{data.frame} with 11 rows and 2 columns
}
\usage{
REV_C08_53
}
\description{
\code{REV_C08_53}
}
\keyword{datasets}
