% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_54}
\alias{REV_C08_54}
\title{Review Exercise 54 of Chapter 8}
\format{
A \link[base]{data.frame} with 22 rows and 8 columns
}
\usage{
REV_C08_54
}
\description{
\code{REV_C08_54}
}
\keyword{datasets}
