% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_57}
\alias{REV_C08_57}
\title{Review Exercise 57 of Chapter 8}
\format{
A \link[base]{data.frame} with 119 rows and 2 columns
}
\usage{
REV_C08_57
}
\description{
\code{REV_C08_57}
}
\keyword{datasets}
