% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_58}
\alias{REV_C08_58}
\title{Review Exercise 58 of Chapter 8}
\format{
A \link[base]{data.frame} with 173 rows and 2 columns
}
\usage{
REV_C08_58
}
\description{
\code{REV_C08_58}
}
\keyword{datasets}
