% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_59}
\alias{REV_C08_59}
\title{Review Exercise 59 of Chapter 8}
\format{
A \link[base]{data.frame} with 192 rows and 2 columns
}
\usage{
REV_C08_59
}
\description{
\code{REV_C08_59}
}
\keyword{datasets}
