% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_62}
\alias{REV_C08_62}
\title{Review Exercise 62 of Chapter 8}
\format{
A \link[base]{data.frame} with 14 rows and 12 columns
}
\usage{
REV_C08_62
}
\description{
\code{REV_C08_62}
}
\keyword{datasets}
