% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C08_64}
\alias{REV_C08_64}
\title{Review Exercise 64 of Chapter 8}
\format{
A \link[base]{data.frame} with 60 rows and 2 columns
}
\usage{
REV_C08_64
}
\description{
\code{REV_C08_64}
}
\keyword{datasets}
