% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C09_21}
\alias{REV_C09_21}
\title{Review Exercise 21 of Chapter 9}
\format{
A \link[base]{data.frame} with 15 rows and 2 columns
}
\usage{
REV_C09_21
}
\description{
\code{REV_C09_21}
}
\keyword{datasets}
