% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C09_32}
\alias{REV_C09_32}
\title{Review Exercise 32 of Chapter 9}
\format{
A \link[base]{data.frame} with 9 rows and 3 columns
}
\usage{
REV_C09_32
}
\description{
\code{REV_C09_32}
}
\keyword{datasets}
