% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C09_34}
\alias{REV_C09_34}
\title{Review Exercise 34 of Chapter 9}
\format{
A \link[base]{data.frame} with 19 rows and 5 columns
}
\usage{
REV_C09_34
}
\description{
\code{REV_C09_34}
}
\keyword{datasets}
