% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C09_40}
\alias{REV_C09_40}
\title{Review Exercise 40 of Chapter 9}
\format{
A \link[base]{data.frame} with 13 rows and 2 columns
}
\usage{
REV_C09_40
}
\description{
\code{REV_C09_40}
}
\keyword{datasets}
