% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C09_43}
\alias{REV_C09_43}
\title{Review Exercise 43 of Chapter 9}
\format{
A \link[base]{data.frame} with 14 rows and 4 columns
}
\usage{
REV_C09_43
}
\description{
\code{REV_C09_43}
}
\keyword{datasets}
