% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C09_44}
\alias{REV_C09_44}
\title{Review Exercise 44 of Chapter 9}
\format{
A \link[base]{data.frame} with 44 rows and 2 columns
}
\usage{
REV_C09_44
}
\description{
\code{REV_C09_44}
}
\keyword{datasets}
