% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C09_46}
\alias{REV_C09_46}
\title{Review Exercise 46 of Chapter 9}
\format{
A \link[base]{data.frame} with 33 rows and 4 columns
}
\usage{
REV_C09_46
}
\description{
\code{REV_C09_46}
}
\keyword{datasets}
