% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C10_06}
\alias{REV_C10_06}
\title{Review Exercise 6 of Chapter 10}
\format{
A \link[base]{data.frame} with 114 rows and 3 columns
}
\usage{
REV_C10_06
}
\description{
\code{REV_C10_06}
}
\keyword{datasets}
