% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C10_07}
\alias{REV_C10_07}
\title{Review Exercise 7 of Chapter 10}
\format{
A \link[base]{data.frame} with 15 rows and 3 columns
}
\usage{
REV_C10_07
}
\description{
\code{REV_C10_07}
}
\keyword{datasets}
