% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C10_11}
\alias{REV_C10_11}
\title{Review Exercise 11 of Chapter 10}
\format{
A \link[base]{data.frame} with 15 rows and 5 columns
}
\usage{
REV_C10_11
}
\description{
\code{REV_C10_11}
}
\keyword{datasets}
