% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C10_17}
\alias{REV_C10_17}
\title{Review Exercise 17 of Chapter 10}
\format{
A \link[base]{data.frame} with 26 rows and 13 columns
}
\usage{
REV_C10_17
}
\description{
\code{REV_C10_17}
}
\keyword{datasets}
