% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C10_18}
\alias{REV_C10_18}
\title{Review Exercise 18 of Chapter 10}
\format{
A \link[base]{data.frame} with 34 rows and 7 columns
}
\usage{
REV_C10_18
}
\description{
\code{REV_C10_18}
}
\keyword{datasets}
