% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C10_19}
\alias{REV_C10_19}
\title{Review Exercise 19 of Chapter 10}
\format{
A \link[base]{data.frame} with 96 rows and 8 columns
}
\usage{
REV_C10_19
}
\description{
\code{REV_C10_19}
}
\keyword{datasets}
