% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C11_14}
\alias{REV_C11_14}
\title{Review Exercise 14 of Chapter 11}
\format{
A \link[base]{data.frame} with 70 rows and 5 columns
}
\usage{
REV_C11_14
}
\description{
\code{REV_C11_14}
}
\keyword{datasets}
