% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C11_22}
\alias{REV_C11_22}
\title{Review Exercise 22 of Chapter 11}
\format{
A \link[base]{data.frame} with 12 rows and 3 columns
}
\usage{
REV_C11_22
}
\description{
\code{REV_C11_22}
}
\keyword{datasets}
