% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C11_28}
\alias{REV_C11_28}
\title{Review Exercise 28 of Chapter 11}
\format{
A \link[base]{data.frame} with 41 rows and 10 columns
}
\usage{
REV_C11_28
}
\description{
\code{REV_C11_28}
}
\keyword{datasets}
