% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C11_29}
\alias{REV_C11_29}
\title{Review Exercise 29 of Chapter 11}
\format{
A \link[base]{data.frame} with 19 rows and 9 columns
}
\usage{
REV_C11_29
}
\description{
\code{REV_C11_29}
}
\keyword{datasets}
