% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C13_06}
\alias{REV_C13_06}
\title{Review Exercise 6 of Chapter 13}
\format{
A \link[base]{data.frame} with 20 rows and 2 columns
}
\usage{
REV_C13_06
}
\description{
\code{REV_C13_06}
}
\keyword{datasets}
