% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C13_09}
\alias{REV_C13_09}
\title{Review Exercise 9 of Chapter 13}
\format{
A \link[base]{data.frame} with 10 rows and 4 columns
}
\usage{
REV_C13_09
}
\description{
\code{REV_C13_09}
}
\keyword{datasets}
