% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C13_10}
\alias{REV_C13_10}
\title{Review Exercise 10 of Chapter 13}
\format{
A \link[base]{data.frame} with 34 rows and 2 columns
}
\usage{
REV_C13_10
}
\description{
\code{REV_C13_10}
}
\keyword{datasets}
