% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C13_16}
\alias{REV_C13_16}
\title{Review Exercise 16 of Chapter 13}
\format{
A \link[base]{data.frame} with 43 rows and 5 columns
}
\usage{
REV_C13_16
}
\description{
\code{REV_C13_16}
}
\keyword{datasets}
