% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C13_17}
\alias{REV_C13_17}
\title{Review Exercise 17 of Chapter 13}
\format{
A \link[base]{data.frame} with 67 rows and 2 columns
}
\usage{
REV_C13_17
}
\description{
\code{REV_C13_17}
}
\keyword{datasets}
