% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C13_20}
\alias{REV_C13_20}
\title{Review Exercise 20 of Chapter 13}
\format{
A \link[base]{data.frame} with 31 rows and 3 columns
}
\usage{
REV_C13_20
}
\description{
\code{REV_C13_20}
}
\keyword{datasets}
