% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C13_21}
\alias{REV_C13_21}
\title{Review Exercise 21 of Chapter 13}
\format{
A \link[base]{data.frame} with 45 rows and 3 columns
}
\usage{
REV_C13_21
}
\description{
\code{REV_C13_21}
}
\keyword{datasets}
