% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C13_23}
\alias{REV_C13_23}
\title{Review Exercise 23 of Chapter 13}
\format{
A \link[base]{data.frame} with 44 rows and 5 columns
}
\usage{
REV_C13_23
}
\description{
\code{REV_C13_23}
}
\keyword{datasets}
