% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C13_24}
\alias{REV_C13_24}
\title{Review Exercise 24 of Chapter 13}
\format{
A \link[base]{data.frame} with 20 rows and 6 columns
}
\usage{
REV_C13_24
}
\description{
\code{REV_C13_24}
}
\keyword{datasets}
