% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C13_27}
\alias{REV_C13_27}
\title{Review Exercise 27 of Chapter 13}
\format{
A \link[base]{data.frame} with 9 rows and 6 columns
}
\usage{
REV_C13_27
}
\description{
\code{REV_C13_27}
}
\keyword{datasets}
