% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C13_29}
\alias{REV_C13_29}
\title{Review Exercise 29 of Chapter 13}
\format{
A \link[base]{data.frame} with 17 rows and 2 columns
}
\usage{
REV_C13_29
}
\description{
\code{REV_C13_29}
}
\keyword{datasets}
