% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{REV_C14_12}
\alias{REV_C14_12}
\title{Review Exercise 12 of Chapter 14}
\format{
A \link[base]{data.frame} with 77 rows and 4 columns
}
\usage{
REV_C14_12
}
\description{
\code{REV_C14_12}
}
\keyword{datasets}
