% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binTab.R
\name{rmd_.binTab}
\alias{rmd_.binTab}
\title{rmd_.binTab}
\usage{
rmd_.binTab(x, xnm = substitute(x), ...)
}
\arguments{
\item{x}{a \link{binTab}}

\item{xnm}{\link[base]{language}, call of \code{x}}

\item{...}{additional parameters, currently not in use}
}
\description{
rmd_.binTab
}
