% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.binTab.R
\name{rmd_binTab}
\alias{rmd_binTab}
\alias{rmd_.binTab}
\alias{Sprintf.binTab}
\title{Create R Markdown Script for \link[DanielBiostatistics10th]{binTab} Object}
\usage{
rmd_.binTab(x, xnm, ...)

Sprintf.binTab(x)
}
\arguments{
\item{x}{a \link[DanielBiostatistics10th]{binTab}}

\item{xnm}{\link[base]{language} or \link[base]{character} scalar, call of \code{x}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=rmd_.binTab]{rmd_.binTab()}} returns a \link[base]{character} \link[base]{vector}.
}
\description{
Create R Markdown Script for \link[DanielBiostatistics10th]{binTab} object.
}
\keyword{internal}
