% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addProbs_vanilla.R
\name{show,addedProbs-method}
\alias{show,addedProbs-method}
\title{Show \linkS4class{addedProbs}}
\usage{
\S4method{show}{addedProbs}(object)
}
\arguments{
\item{object}{\linkS4class{addedProbs} object}
}
\value{
The \link[methods]{show} method for \linkS4class{addedProbs} object
does not have a returned value.
}
\description{
..
}
