% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0S4_Boolean.R
\name{show,BooleanRisk-method}
\alias{show,BooleanRisk-method}
\title{Show \linkS4class{BooleanRisk} Object}
\usage{
\S4method{show}{BooleanRisk}(object)
}
\arguments{
\item{object}{a \linkS4class{BooleanRisk} object}
}
\value{
The \link[methods]{show} method for \linkS4class{BooleanRisk} object 
does not have a returned value.
}
\description{
Show \linkS4class{BooleanRisk} object
}
