% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BooleanTable.R
\name{show,BooleanTable-method}
\alias{show,BooleanTable-method}
\title{Show \linkS4class{BooleanTable} Object}
\usage{
\S4method{show}{BooleanTable}(object)
}
\arguments{
\item{object}{a \linkS4class{BooleanTable} object}
}
\value{
The \link[methods]{show} method for \linkS4class{BooleanTable} object
does not have a returned value.
}
\description{
Show \linkS4class{BooleanTable} object
}
\keyword{internal}
