% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0S4_Boolean.R
\name{show,BooleanTest-method}
\alias{show,BooleanTest-method}
\title{Show \linkS4class{BooleanTest} Object}
\usage{
\S4method{show}{BooleanTest}(object)
}
\arguments{
\item{object}{a \linkS4class{BooleanTest} object}
}
\value{
The \link[methods]{show} method for \linkS4class{BooleanTest} object 
does not have a returned value.
}
\description{
Show \linkS4class{BooleanTest} object
}
