% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chp2.R
\name{show,freqs-method}
\alias{show,freqs-method}
\title{Show \linkS4class{freqs} Object}
\usage{
\S4method{show}{freqs}(object)
}
\arguments{
\item{object}{an \linkS4class{freqs} object}
}
\value{
The \link[methods]{show} method for \linkS4class{freqs} object 
does not have a returned value.
}
\description{
Show \linkS4class{freqs} object
}
