% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chp7.power.R
\name{show,power-method}
\alias{show,power-method}
\title{Show \linkS4class{power} Object}
\usage{
\S4method{show}{power}(object)
}
\arguments{
\item{object}{a \linkS4class{power} object}
}
\value{
The \link[methods]{show} method for \linkS4class{power} object 
does not have a returned value.
}
\description{
Show \linkS4class{power} object
}
