% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kappa.R
\name{Sprintf.Kappa}
\alias{Sprintf.Kappa}
\title{Summary Information of Cohen's \link[vcd]{Kappa}}
\usage{
Sprintf.Kappa(model, ...)
}
\arguments{
\item{model}{a \link[vcd]{Kappa} object}

\item{...}{additional parameters, currently not in use}
}
\description{
Summary Information of Cohen's \link[vcd]{Kappa}
}
\examples{
# ?vcd::Kappa
library(vcd)
data('SexualFun', package = 'vcd')
cat(Sprintf.Kappa(Kappa(SexualFun)))
}
\keyword{internal}
