% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{\eqn{t}-test from Aggregated Statistics}
\usage{
t_test(
  xbar,
  xsd,
  n,
  null.value,
  var.equal = FALSE,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{xbar}{\link[base]{numeric} scalar \eqn{\bar{x}} or
\link[base]{length}-2 \link[base]{vector} \eqn{(\bar{x}_1, \bar{x}_2)}.
In the case of two-sample tests, this could also be a \link[base]{numeric} scalar \eqn{\bar{x}_1-\bar{x}_2}.}

\item{xsd}{\link[base]{numeric} scalar \eqn{s} or
\link[base]{length}-2 \link[base]{vector} \eqn{(s_1, s_2)},
sample standard deviation(s)}

\item{n}{\link[base]{integer} scalar \eqn{n} or
\link[base]{length}-2 \link[base]{vector} \eqn{(n_1, n_2)},
sample size(s)}

\item{null.value}{(optional) \link[base]{numeric} scalar or \link[base]{length}-2 \link[base]{vector}.
Null value(s) of the population mean(s)
(\eqn{\mu_0}, \eqn{(\mu_{10}, \mu_{20})}, or \eqn{\mu_{10}-\mu_{20}}).
If missing, only the confidence intervals will be computed.}

\item{var.equal}{\link[base]{logical} scalar, whether to treat the two population variances as being equal
(default \code{FALSE}) in function \link{t_test}}

\item{alternative}{\link[base]{character} scalar, alternative hypothesis,
either \code{'two.sided'} (default), \code{'greater'} or \code{'less'}}

\item{conf.level}{\link[base]{numeric} scalar \eqn{(1-\alpha)}, confidence level, default 0.95}

\item{...}{potential arguments, not in use currently}
}
\value{
Function \code{\link[=t_test]{t_test()}} returns an \link[stats:t.test]{htest} object when \code{null.value} is provided,
otherwise returns a length-two \link[base]{numeric} vector.
}
\description{
\eqn{t}-test from aggregated statistics.
}
\details{
Function \code{\link[=t_test]{t_test()}} performs one- or two-sample \eqn{t}-test
using the aggregated statistics of sample mean(s), sample standard deviation(s) and sample size(s)
when \code{null.value} is provided.  Otherwise, only the confidence interval based on
\eqn{t}-distribution is computed.
}
\seealso{
\link[stats]{t.test}
}
\keyword{internal}
