% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_test.R
\name{var_test}
\alias{var_test}
\title{One-Sample \eqn{\chi^2}-test,
or Two-Sample \eqn{F}-test, on Variances}
\usage{
var_test(
  xsd = sd(obs),
  n = length(obs),
  obs,
  null.value,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{xsd}{\link[base]{numeric} scalar \eqn{s} or
\link[base]{length}-2 \link[base]{vector} \eqn{(s_1, s_2)},
sample standard deviation(s)}

\item{n}{\link[base]{integer} scalar \eqn{n} or
\link[base]{length}-2 \link[base]{vector} \eqn{(n_1, n_2)},
sample size(s)}

\item{obs}{\link[base]{numeric} \link[base]{vector}, observations}

\item{null.value}{(optional) \link[base]{numeric} scalar or \link[base]{length}-2 \link[base]{vector}.
Null value(s) of the population variance(s) (ratio)
(\eqn{\sigma^2_0}, \eqn{(\sigma^2_{10}, \sigma^2_{20})}, or \eqn{\sigma^2_{10}/\sigma^2_{20}}).
If missing, only the confidence intervals will be computed.}

\item{alternative}{\link[base]{character} scalar, alternative hypothesis,
either \code{'two.sided'} (default), \code{'greater'} or \code{'less'}}

\item{conf.level}{\link[base]{numeric} scalar \eqn{(1-\alpha)}, confidence level, default 0.95}

\item{...}{potential arguments, currently not in use}
}
\value{
Function \code{\link[=var_test]{var_test()}} returns an \link[stats:var.test]{htest} object when \code{null.value} is provided,
otherwise returns a length-two \link[base]{numeric} vector.
}
\description{
One-sample \eqn{\chi^2}-test,
or two-sample \eqn{F}-test, on variances.
}
\details{
Function \code{\link[=var_test]{var_test()}} performs one-sample \eqn{\chi^2}-test on variance,
or two-sample \eqn{F}-test on variances, using the aggregated statistics of
sample standard deviation(s) and sample size(s) when \code{null.value} is provided.
Otherwise, only the confidence interval based on \eqn{\chi^2}- or \eqn{F}-distribution is computed.
}
\keyword{internal}
