% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chp5.6.7.R
\name{Chapter05to07}
\alias{Chapter05to07}
\alias{aggregated_z}
\alias{aggregated_t}
\alias{prop_CLT}
\alias{aggregated_var}
\title{Chapter 5, 6 and 7}
\usage{
aggregated_z(
  xbar,
  n,
  sd,
  null.value,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  ...
)

aggregated_t(
  xbar,
  xsd,
  n,
  null.value,
  var.equal = FALSE,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  ...
)

prop_CLT(
  x,
  n,
  obs,
  xbar = x/n,
  null.value,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  ...
)

aggregated_var(
  xsd,
  n,
  null.value,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{xbar}{\link[base]{numeric} scalar or
\link[base]{length}-2 \link[base]{vector}.
Sample mean(s) for \link[base]{numeric} variable(s) \eqn{\bar{x}} or \eqn{(\bar{x}_1, \bar{x}_2)}.
Sample proportion(s) for binary (i.e., \link[base]{logical}) variable(s) \eqn{\hat{p}} or \eqn{(\hat{p}_1, \hat{p}_2)}.
In the case of two-sample tests, this could also be a \link[base]{numeric} scalar indicating the difference in
sample means \eqn{\bar{x}_1-\bar{x}_2} or sample proportions \eqn{\hat{p}_1-\hat{p}_2}}

\item{n}{\link[base]{integer} scalar \eqn{n} or
\link[base]{length}-2 \link[base]{vector} \eqn{(n_1, n_2)},
sample size(s)}

\item{sd}{\link[base]{numeric} scalar \eqn{\sigma} or
\link[base]{length}-2 \link[base]{vector} \eqn{(\sigma_1, \sigma_2)},
population standard deviation(s)}

\item{null.value}{(optional) \link[base]{numeric} scalar or \link[base]{length}-2 \link[base]{vector}.
Null value(s) of the population mean(s)
(\eqn{\mu_0}, \eqn{(\mu_{10}, \mu_{20})}, or \eqn{\mu_{10}-\mu_{20}})
for functions \link{aggregated_z} and \link{aggregated_t}.
Null value(s) of the population proportion(s)
(\eqn{p_0}, \eqn{(p_{10}, p_{20})}, or \eqn{p_{10}-p_{20}})
for \link{prop_CLT}.
Null value(s) of the population variance(s) (ratio)
(\eqn{\sigma^2_0}, \eqn{(\sigma^2_{10}, \sigma^2_{20})}, or \eqn{\sigma^2_{10}/\sigma^2_{20}})
for function \link{aggregated_var}.
If missing, only the confidence intervals will be computed.}

\item{alternative}{\link[base]{character} scalar, alternative hypothesis,
either \code{'two.sided'} (default), \code{'greater'} or \code{'less'}}

\item{conf.level}{\link[base]{numeric} scalar \eqn{(1-\alpha)}, confidence level, default 0.95}

\item{...}{potential arguments, not in use currently}

\item{xsd}{\link[base]{numeric} scalar \eqn{s} or
\link[base]{length}-2 \link[base]{vector} \eqn{(s_1, s_2)},
sample standard deviation(s)}

\item{var.equal}{\link[base]{logical} scalar, whether to treat the two population variances as being equal
(default \code{FALSE}) in function \link{aggregated_t}}

\item{x}{\link[base]{integer} scalar or \link[base]{length}-2 \link[base]{vector},
number of positive count(s) of binary (i.e., \link[base]{logical}) variable(s)}

\item{obs}{\link[base]{vector}, observations,
currently used only in one-sample \eqn{z}-test on proportion \link{prop_CLT}}
}
\value{
Function \code{\link[=aggregated_z]{aggregated_z()}} returns an \code{'htest'} object when \code{null.value} is provided,
otherwise returns a length-two \link[base]{numeric} vector.

Function \code{\link[=aggregated_t]{aggregated_t()}} returns an \link[stats:t.test]{htest} object when \code{null.value} is provided,
otherwise returns a length-two \link[base]{numeric} vector.

Function \code{\link[=prop_CLT]{prop_CLT()}} returns an \link[stats:prop.test]{htest} object when \code{null.value} is provided,
otherwise returns a length-two \link[base]{numeric} vector.

Function \code{\link[=aggregated_var]{aggregated_var()}} returns an \link[stats:var.test]{htest} object when \code{null.value} is provided,
otherwise returns a length-two \link[base]{numeric} vector.
}
\description{
Functions for Chapter 5, \emph{Some Important Sampling Distributions},
Chapter 6, \emph{Estimation} and
Chapter 7, \emph{Hypothesis Testing}.
}
\details{
Function \code{\link[=aggregated_z]{aggregated_z()}} performs one- or two-sample \eqn{z}-test
using the aggregated statistics of sample mean(s) and sample size(s) when
\code{null.value} is provided.  Otherwise, only the confidence interval based on
\eqn{z}-distribution is computed.

Function \code{\link[=aggregated_t]{aggregated_t()}} performs one- or two-sample \eqn{t}-test
using the aggregated statistics of sample mean(s), sample standard deviation(s) and sample size(s)
when \code{null.value} is provided.  Otherwise, only the confidence interval based on
\eqn{t}-distribution is computed.

Function \code{\link[=prop_CLT]{prop_CLT()}} performs one- or two-sample \eqn{z}-test on proportion(s),
using Central Limit Theorem when \code{null.value} is provided.
Otherwise, only the confidence interval based on \eqn{z}-distribution is computed.

Function \code{\link[=aggregated_var]{aggregated_var()}} performs one-sample \eqn{\chi^2}-test on variance,
or two-sample \eqn{F}-test on variances, using the aggregated statistics of
sample standard deviation(s) and sample size(s) when \code{null.value} is provided.
Otherwise, only the confidence interval based on \eqn{\chi^2}- or \eqn{F}-distribution is computed.
}
\seealso{
\link[stats]{t.test} \link[stats]{prop.test} \link[stats]{var.test}
}
\keyword{internal}
