% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DanielBiostatistics10th.R
\docType{package}
\name{DanielBiostatistics10th-package}
\alias{DanielBiostatistics10th-package}
\title{Functions for Wayne W. Daniel's \emph{Biostatistics} (Tenth Edition)}
\description{
R functions to accompany Wayne W. Daniel's 
\emph{Biostatistics: A Foundation for Analysis in the Health Sciences}, Tenth Edition,
for Course \code{GC660} of Jefferson Medical College, Thomas Jefferson University, Philadelphia, PA.
}
\references{
Wayne W. Daniel, \emph{Biostatistics: A Foundation for Analysis in the Health Sciences}, Tenth Edition.
Wiley, ISBN: 978-1-119-62550-6.
}
\keyword{package}
