% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DanielBiostatistics10th.R
\docType{package}
\name{DanielBiostatistics10th-package}
\alias{DanielBiostatistics10th-package}
\title{Functions for Wayne W. Daniel's \emph{Biostatistics} (Tenth Edition)}
\description{
Functions and examples to accompany Wayne W. Daniel's 
\emph{Biostatistics: A Foundation for Analysis in the Health Sciences}, Tenth Edition.
}
\references{
Wayne W. Daniel, \emph{Biostatistics: A Foundation for Analysis in the Health Sciences}, Tenth Edition.
Wiley, ISBN: 978-1-119-62550-6.
}
\keyword{package}
