% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binTab.R
\name{predval}
\alias{predval}
\alias{ppv}
\alias{npv}
\title{Predictive Values}
\usage{
ppv(prevalence, sensitivity, specificity)

npv(prevalence, sensitivity, specificity)
}
\arguments{
\item{prevalence}{\link[base]{double} scalar or \link[base]{vector}}

\item{sensitivity, specificity}{\link[base]{double} scalars}
}
\value{
Functions \code{\link[=ppv]{ppv()}} and \code{\link[=npv]{npv()}} return \link[base]{double} scalar or \link[base]{vector}.
}
\description{
Positive and negative predictive values.
}
\details{
Function \code{\link[=ppv]{ppv()}} calculates positive predictive values.

Function \code{\link[=npv]{npv()}} calculates negative predictive values.
}
\keyword{internal}
