% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binTab.R
\name{print.binTab}
\alias{print.binTab}
\title{Print Boolean Test-&-Disease and/or Risk-&-Disease Table}
\usage{
\method{print}{binTab}(x, ...)
}
\arguments{
\item{x}{a \link{binTab} object}

\item{...}{potential parameters of function \link{Sprintf.binTab}}
}
\value{
Function \link{print.binTab} does not have a returned value.
}
\description{
Print Boolean test-&-disease and/or risk-&-disease table.
}
\note{
Function \link[e1071]{classAgreement} does not provide confidence interval of \eqn{\kappa}.
}
\examples{
(x = array(c(95L, 10L, 31L, 82L), dim = c(2L, 2L)))
binTab(x)
print(binTab(x), prevalence = c(.0001, .001, .01))
}
\references{
\url{https://en.wikipedia.org/wiki/Diagnostic_odds_ratio}
}
\keyword{internal}
