% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_CLT.R
\name{prop_CLT}
\alias{prop_CLT}
\title{\eqn{z}-test on Proportion, using Central Limit Theorem}
\usage{
prop_CLT(
  x,
  n,
  obs,
  phat = x/n,
  null.value,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{x}{\link[base]{integer} scalar or \link[base]{length}-2 \link[base]{vector},
number of positive count(s) of binary (i.e., \link[base]{logical}) variable(s)}

\item{n}{\link[base]{integer} scalar \eqn{n} or
\link[base]{length}-2 \link[base]{vector} \eqn{(n_1, n_2)},
sample size(s)}

\item{obs}{\link[base]{logical} \link[base]{vector}, observations}

\item{phat}{\link[base]{numeric} scalar \eqn{\hat{p}} or
\link[base]{length}-2 \link[base]{vector} \eqn{(\hat{p}_1, \hat{p}_2)}.
In the case of two-sample tests, this could also be a \link[base]{numeric} scalar \eqn{\hat{p}_1-\hat{p}_2}}

\item{null.value}{(optional) \link[base]{numeric} scalar or \link[base]{length}-2 \link[base]{vector}.
Null value(s) of the population proportion(s)
(\eqn{p_0}, \eqn{(p_{10}, p_{20})}, or \eqn{p_{10}-p_{20}})
for function \code{\link[=prop_CLT]{prop_CLT()}}.
If missing, only the confidence intervals will be computed.}

\item{alternative}{\link[base]{character} scalar, alternative hypothesis,
either \code{'two.sided'} (default), \code{'greater'} or \code{'less'}}

\item{conf.level}{\link[base]{numeric} scalar \eqn{(1-\alpha)}, confidence level, default 0.95}

\item{...}{potential arguments, not in use currently}
}
\value{
Function \code{\link[=prop_CLT]{prop_CLT()}} returns an \link[stats:prop.test]{htest} object when \code{null.value} is provided,
otherwise returns a length-two \link[base]{numeric} vector.
}
\description{
\eqn{z}-test on proportion, using Central Limit Theorem.
}
\details{
Function \code{\link[=prop_CLT]{prop_CLT()}} performs one- or two-sample \eqn{z}-test on proportion(s),
using Central Limit Theorem when \code{null.value} is provided.
Otherwise, only the confidence interval based on \eqn{z}-distribution is computed.
}
\seealso{
\link[stats]{prop.test}
}
\keyword{internal}
