% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chp1_sampleRow.R
\name{sampleRow}
\alias{sampleRow}
\title{Simple Random Sample from a \link[base]{data.frame}}
\usage{
sampleRow(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{a \link[base]{data.frame}}

\item{size}{positive \link[base]{integer} scalar, number of rows to be selected}

\item{replace}{\link[base]{logical} scalar, whether sampling should be with replacement (default \code{FALSE})}

\item{prob}{\link[base]{numeric} vector of probability weights for each row of input \code{x} being sampled.
Default \code{NULL} indicates simple random sampling}
}
\value{
Function \link{sampleRow} returns a \link[base]{data.frame}, a simple random sample from the input.
}
\description{
Functions and examples for Chapter 1, \emph{Introduction to Biostatistics}.
}
\seealso{
\link[base]{sample.int}
}
\keyword{internal}
