% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binTab.R
\name{Sprintf.binTab}
\alias{Sprintf.binTab}
\title{Printout for \link{binTab} Object}
\usage{
Sprintf.binTab(model, prevalence, ...)
}
\arguments{
\item{model}{a \link{binTab} object}

\item{prevalence}{(optional) \link[base]{numeric} scalar or \link[base]{vector}, prevalence of disease}

\item{...}{potential parameters, currently not in use}
}
\value{
Function \link{Sprintf.binTab} returns a \link[base]{character} \link[base]{vector}.
}
\description{
Printout for \link{binTab} Object
}
\details{
Function \link{Sprintf.binTab} prints the diagnostic test characteristics,
e.g., sensitivity, specificity, predictive values, and diagnostic accuracy,
together with their \eqn{95\%} Clopper-Pearson exact confidence intervals.
}
\keyword{internal}
