% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chp3_binTab.R
\name{Sprintf.binTab}
\alias{Sprintf.binTab}
\title{Description for \link[DanielBiostatistics10th]{binTab} Object}
\usage{
Sprintf.binTab(x)
}
\arguments{
\item{x}{a \link[DanielBiostatistics10th]{binTab}}
}
\value{
Function \code{\link[=Sprintf.binTab]{Sprintf.binTab()}} returns a \link[base]{character} scalar.
}
\description{
description for \link[DanielBiostatistics10th]{binTab} object.
}
\keyword{internal}
