% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binTab.R
\name{summary.binTab}
\alias{summary.binTab}
\title{Summarize Boolean Test-&-Disease and/or Risk-&-Disease Table}
\usage{
\method{summary}{binTab}(object, prevalence, ...)
}
\arguments{
\item{object}{a \link{binTab}}

\item{prevalence}{(optional) \link[base]{numeric} scalar, prevalence of disease}

\item{...}{potential parameters, currently not in use}
}
\value{
Function \link{summary.binTab} returns the sensitivity and specificity invisibly.
}
\description{
Summarize Boolean test-&-disease and/or risk-&-disease table using
sensitivity, specificity, diagnostic accuracy, predictive values, relative risk
and odds ratio,
together with their \eqn{95\%} Clopper-Pearson exact confidence intervals.
}
\details{
..
}
\examples{
(x = array(c(95L, 10L, 31L, 82L), dim = c(2L, 2L)))
summary(binTab(x))
summary(binTab(x), prevalence = c(.0001, .001, .01))

}
\references{
\url{https://en.wikipedia.org/wiki/Diagnostic_odds_ratio}
}
\keyword{internal}
