% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chp3_binTab.R
\name{summary.binTab}
\alias{summary.binTab}
\title{Print Boolean Test-&-Disease and/or Risk-&-Disease Table}
\usage{
\method{summary}{binTab}(object, prevalence, ...)
}
\arguments{
\item{object}{a \link{binTab} object}

\item{prevalence}{(optional) \link[base]{numeric} scalar or \link[base]{vector}, prevalence of disease}

\item{...}{potential parameters, currently not in use}
}
\value{
Function \link{summary.binTab} \link[base]{invisible}-y returns a \link[base]{character} \link[base]{vector}.
}
\description{
Print Boolean test-&-disease and/or risk-&-disease table.
}
\details{
Function \link{summary.binTab} prints the diagnostic test characteristics,
e.g., sensitivity, specificity, predictive values, and diagnostic accuracy,
together with their \eqn{95\%} Clopper-Pearson exact confidence intervals.
}
\note{
Function \link[e1071]{classAgreement} does not provide confidence interval of \eqn{\kappa}.
}
\examples{
(x = array(c(95L, 10L, 31L, 82L), dim = c(2L, 2L)))
binTab(x)
print(binTab(x), prevalence = c(.0001, .001, .01))
}
\references{
\url{https://en.wikipedia.org/wiki/Diagnostic_odds_ratio}
}
\keyword{internal}
