% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewOE.R
\name{update_df}
\alias{update_df}
\title{Additional Restriction on \link[stats]{chisq.test}}
\usage{
update_df(x, restriction)
}
\arguments{
\item{x}{returned object of function \link[stats]{chisq.test}}

\item{restriction}{\link[base]{integer} scalar,
number of additional restrictions.
Or, \link[base]{character} \link[base]{vector},
brief explanation of each additional restriction.}
}
\description{
Additional Restriction on \link[stats]{chisq.test}
}
\keyword{internal}
