% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chp2_viewStats_viewFreq.R
\name{viewFreq}
\alias{viewFreq}
\title{View Frequency Table}
\usage{
viewFreq(x)
}
\arguments{
\item{x}{an \link[base]{ordered} \link[base]{factor}}
}
\value{
Function \code{\link[=viewFreq]{viewFreq()}} returns a \link[flextable]{flextable}
}
\description{
View Frequency Table
}
\details{
Function \code{\link[=viewFreq]{viewFreq()}} shows the (relative) frequencies and cumulative (relative) frequencies.
}
\keyword{internal}
