% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewOE.R
\name{viewOE}
\alias{viewOE}
\title{View Observed and Expected Frequency Table}
\usage{
viewOE(x)
}
\arguments{
\item{x}{returned object of function \link[stats]{chisq.test}}
}
\value{
Function \code{\link[=viewOE]{viewOE()}} returns a \link[flextable]{flextable} of
observed and expected frequencies, as well as
the category-wise \eqn{\chi^2} statistics.
}
\description{
View Observed and Expected Frequency Table
}
\keyword{internal}
