% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chp2_viewStats_viewFreq.R
\name{viewStats}
\alias{viewStats}
\title{Print Summary Statistics}
\usage{
viewStats(x, na.rm = TRUE)
}
\arguments{
\item{x}{\link[base]{numeric} vector, the observations.}

\item{na.rm}{\link[base]{logical} scalar, whether to remove the missing observations, default \code{TRUE}.}
}
\value{
Function \code{\link[=viewStats]{viewStats()}} does not have a returned value.
}
\description{
Print summary statistics.
}
\details{
Function \code{\link[=viewStats]{viewStats()}} prints the simple statistics of the input observations, such as sample size,
mean, median, (smallest) mode, variance, standard deviation,
coefficient of variation (if all observations are non-negative),
quartiles, inter-quartile range (IQR), range, skewness and kurtosis.  A histogram is also printed.
}
\keyword{internal}
