% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_test.R
\name{z_test}
\alias{z_test}
\title{\eqn{z}-test from Aggregated Statistics}
\usage{
z_test(
  xbar = mean(obs),
  n = length(obs),
  obs,
  sd,
  null.value,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{xbar}{\link[base]{numeric} scalar \eqn{\bar{x}} or
\link[base]{length}-2 \link[base]{vector} \eqn{(\bar{x}_1, \bar{x}_2)}.
In the case of two-sample tests, this could also be a \link[base]{numeric} scalar \eqn{\bar{x}_1-\bar{x}_2}.}

\item{n}{\link[base]{integer} scalar \eqn{n} or
\link[base]{length}-2 \link[base]{vector} \eqn{(n_1, n_2)},
sample size(s)}

\item{obs}{\link[base]{numeric} \link[base]{vector}, observations}

\item{sd}{\link[base]{numeric} scalar \eqn{\sigma} or
\link[base]{length}-2 \link[base]{vector} \eqn{(\sigma_1, \sigma_2)},
population standard deviation(s)}

\item{null.value}{(optional) \link[base]{numeric} scalar or \link[base]{length}-2 \link[base]{vector}.
Null value(s) of the population mean(s)
(\eqn{\mu_0}, \eqn{(\mu_{10}, \mu_{20})}, or \eqn{\mu_{10}-\mu_{20}}).
If missing, only the confidence intervals will be computed.}

\item{alternative}{\link[base]{character} scalar, alternative hypothesis,
\code{'two.sided'} (default), \code{'greater'} or \code{'less'}}

\item{conf.level}{\link[base]{numeric} scalar \eqn{(1-\alpha)}, confidence level, default 0.95}

\item{...}{potential arguments, currently not in use}
}
\value{
Function \code{\link[=z_test]{z_test()}} returns an \code{'htest'} object when \code{null.value} is provided,
otherwise returns a length-two \link[base]{numeric} vector.
}
\description{
\eqn{z}-test from aggregated statistics
}
\details{
Function \code{\link[=z_test]{z_test()}} performs one- or two-sample \eqn{z}-test
using the aggregated statistics of sample mean(s) and sample size(s) when
\code{null.value} is provided.  Otherwise, only the confidence interval based on
\eqn{z}-distribution is computed.
}
\keyword{internal}
